move deckManaCurve and deckTypesStats to shared util
