+++ shared/util.js
@@ -1,14 +1,17 @@
-const db = require("../shared/database");
-const pd = require("../shared/player-data");
-
 const {
   FORMATS,
   BLACK,
   BLUE,
   GREEN,
   RED,
-  WHITE
-} = require("../shared/constants.js");
+  WHITE,
+  MANA_COLORS,
+  CARD_TYPES,
+  CARD_TYPE_CODES
+} = require("../shared/constants");
+const db = require("../shared/database");
+const pd = require("../shared/player-data");
+const { createDiv, createSpan } = require("../shared/dom-fns");
 
 //
 exports.get_card_image = get_card_image;
@@ -384,10 +387,17 @@ function getBoosterCountEstimate(wildcards) {
 //
 exports.get_deck_types_ammount = get_deck_types_ammount;
 function get_deck_types_ammount(deck) {
-  var types = { art: 0, cre: 0, enc: 0, ins: 0, lan: 0, pla: 0, sor: 0 };
+  const types = { art: 0, cre: 0, enc: 0, ins: 0, lan: 0, pla: 0, sor: 0 };
+  if (!deck.mainDeck) return types;
 
   deck.mainDeck.forEach(function(card) {
-    var c = db.card(card.id);
+    // This is hackish.. the way we insert our custom elements in the
+    // array of cards is wrong in the first place :()
+    if (card.id.id && card.id.id == 100) {
+      types.lan += card.quantity;
+      return;
+    }
+    const c = db.card(card.id);
     if (c) {
       if (c.type.includes("Land", 0)) types.lan += card.quantity;
       else if (c.type.includes("Creature", 0)) types.cre += card.quantity;
@@ -400,6 +410,48 @@ function get_deck_types_ammount(deck) {
   });
 
   return types;
+}
+
+//
+exports.get_deck_curve = get_deck_curve;
+function get_deck_curve(deck) {
+  const curve = [];
+  if (!deck.mainDeck) return curve;
+
+  deck.mainDeck.forEach(card => {
+    const cardObj = db.card(card.id);
+    if (!cardObj) return;
+
+    const cmc = cardObj.cmc;
+    if (!curve[cmc]) curve[cmc] = [0, 0, 0, 0, 0, 0];
+
+    if (!cardObj.type.includes("Land")) {
+      cardObj.cost.forEach(c => {
+        if (c.includes("w")) curve[cmc][1] += card.quantity;
+        if (c.includes("u")) curve[cmc][2] += card.quantity;
+        if (c.includes("b")) curve[cmc][3] += card.quantity;
+        if (c.includes("r")) curve[cmc][4] += card.quantity;
+        if (c.includes("g")) curve[cmc][5] += card.quantity;
+      });
+
+      curve[cmc][0] += card.quantity;
+    }
+  });
+  /*
+  // Do not account sideboard?
+  deck.sideboard.forEach(function(card) {
+    var grpid = card.id;
+    var cmc = db.card(grpid).cmc;
+    if (curve[cmc] == undefined)  curve[cmc] = 0;
+    curve[cmc] += card.quantity
+
+    if (db.card(grpid).rarity !== 'land') {
+      curve[cmc] += card.quantity
+    }
+  });
+  */
+  //console.log(curve);
+  return curve;
 }
 
 //
@@ -645,4 +697,74 @@ function add(a, b) {
 exports.objectClone = objectClone;
 function objectClone(originalObject) {
   return JSON.parse(JSON.stringify(originalObject));
+}
+
+//
+exports.deckManaCurve = deckManaCurve;
+function deckManaCurve(deck) {
+  const manaCounts = get_deck_curve(deck);
+  const curveMax = Math.max(
+    ...manaCounts
+      .filter(v => {
+        if (v == undefined) return false;
+        return true;
+      })
+      .map(v => v[0] || 0)
+  );
+  // console.log("deckManaCurve", manaCounts, curveMax);
+
+  const container = createDiv(["mana_curve_container"]);
+  const curve = createDiv(["mana_curve"]);
+  const numbers = createDiv(["mana_curve_numbers"]);
+
+  manaCounts.forEach((cost, i) => {
+    const total = cost[0];
+    const manaTotal = cost.reduce(add, 0) - total;
+
+    const curveCol = createDiv(["mana_curve_column"]);
+    curveCol.style.height = (total * 100) / curveMax + "%";
+
+    const curveNum = createDiv(["mana_curve_number"], total > 0 ? total : "");
+    curveCol.appendChild(curveNum);
+
+    MANA_COLORS.forEach((mc, ind) => {
+      if (ind < 5 && cost[ind + 1] > 0) {
+        const col = createDiv(["mana_curve_column_color"]);
+        col.style.height = Math.round((cost[ind + 1] / manaTotal) * 100) + "%";
+        col.style.backgroundColor = mc;
+        curveCol.appendChild(col);
+      }
+    });
+
+    curve.appendChild(curveCol);
+
+    const colNum = createDiv(["mana_curve_column_number"]);
+    const numDiv = createDiv(["mana_s16", "mana_" + i]);
+    numDiv.style.margin = "0 auto !important";
+    colNum.appendChild(numDiv);
+    numbers.appendChild(colNum);
+  });
+
+  container.appendChild(curve);
+  container.appendChild(numbers);
+
+  return container;
+}
+
+//
+exports.deckTypesStats = deckTypesStats;
+function deckTypesStats(deck) {
+  const cardTypes = get_deck_types_ammount(deck);
+  const typesContainer = createDiv(["types_container"]);
+  CARD_TYPE_CODES.forEach((cardTypeKey, index) => {
+    const type = createDiv(["type_icon_cont"]);
+    type.appendChild(
+      createDiv(["type_icon", "type_" + cardTypeKey], "", {
+        title: CARD_TYPES[index]
+      })
+    );
+    type.appendChild(createSpan([], cardTypes[cardTypeKey]));
+    typesContainer.appendChild(type);
+  });
+  return typesContainer;
 }
